Removed search model and moved the dataclass definitions to extension_interface
